Enable bugprone-misplaced-widening-cast (#59534)
This clang-tidy check looks for operations where an arithmetic operation
is cast to a larger result.  E.g. static_cast<long>( a + b ) when a and
b are ints.  The expectation is that such casts are misplaced because if
a + b overflows, you would rather have cast it to a long before adding
b.

We had a few hits for this warning.  I don't think any of them were real
issues; mostly just casts for the sake of avoiding signed/unsigned
comparison warnings.  But I refactored them anyway.

+++ src/map.h
@@ -354,13 +354,13 @@ class map
         bool check_seen_cache( const tripoint &p ) const {
             std::bitset<MAPSIZE_X *MAPSIZE_Y> &memory_seen_cache =
                 get_cache( p.z ).map_memory_seen_cache;
-            return !memory_seen_cache[ static_cast<size_t>( p.x + p.y * MAPSIZE_Y ) ];
+            return !memory_seen_cache[ p.x + p.y * MAPSIZE_Y ];
         }
         bool check_and_set_seen_cache( const tripoint &p ) const {
             std::bitset<MAPSIZE_X *MAPSIZE_Y> &memory_seen_cache =
                 get_cache( p.z ).map_memory_seen_cache;
-            if( !memory_seen_cache[ static_cast<size_t>( p.x + p.y * MAPSIZE_Y ) ] ) {
-                memory_seen_cache.set( static_cast<size_t>( p.x + p.y * MAPSIZE_Y ) );
+            if( !memory_seen_cache[ p.x + p.y * MAPSIZE_Y ] ) {
+                memory_seen_cache.set( p.x + p.y * MAPSIZE_Y );
                 return true;
             }
             return false;
